Implement netgame input logic Fix a crash in Linux with OGG audio by using stb_vorbis Fix a few warnings and errors under gcc
The netgame logic uses the lockstep model.
There is no guarantee that a netgame will be deterministic.
The netcode will not attempt to synchronize the state of a Scene,
such as its Entities; or the RNG state, which is not seeded in a
deterministic way.

This is a work in progress.